Parametric Bootstrap for the Linear Stochastic Differential Equation Model using a State Space Model Parameterization (Fixed Parameters)

This function simulates data from a linear stochastic differential equation model using a state-space model parameterization and fits the model using the dynr package. The process is repeated R times. It assumes that the parameters remain constant across individuals and over time. At the moment, the function only supports type = 0.

Usage

PBSSMLinSDEFixed(
  R,
  path,
  prefix,
  n,
  time,
  delta_t = 0.1,
  mu0,
  sigma0_l,
  iota,
  phi,
  sigma_l,
  nu,
  lambda,
  theta_l,
  type = 0,
  x = NULL,
  gamma = NULL,
  kappa = NULL,
  mu0_fixed = FALSE,
  sigma0_fixed = FALSE,
  alpha_level = 0.05,
  optimization_flag = TRUE,
  hessian_flag = FALSE,
  verbose = FALSE,
  weight_flag = FALSE,
  debug_flag = FALSE,
  perturb_flag = FALSE,
  xtol_rel = 1e-07,
  stopval = -9999,
  ftol_rel = -1,
  ftol_abs = -1,
  maxeval = as.integer(-1),
  maxtime = -1,
  ncores = NULL,
  seed = NULL,
  clean = TRUE
)

Arguments

R

Positive integer. Number of bootstrap samples.

path

Path to a directory to store bootstrap samples and estimates.

prefix

Character string. Prefix used for the file names for the bootstrap samples and estimates.

n

Positive integer. Number of individuals.

time

Positive integer. Number of time points.

delta_t

Numeric. Time interval (\(\Delta_t\)).

mu0

Numeric vector. Mean of initial latent variable values (\(\boldsymbol{\mu}_{\boldsymbol{\eta} \mid 0}\)).

sigma0_l

Numeric matrix. Cholesky factorization (t(chol(sigma0))) of the covariance matrix of initial latent variable values (\(\boldsymbol{\Sigma}_{\boldsymbol{\eta} \mid 0}\)).

iota

Numeric vector. An unobserved term that is constant over time (\(\boldsymbol{\iota}\)).

phi

Numeric matrix. The drift matrix which represents the rate of change of the solution in the absence of any random fluctuations (\(\boldsymbol{\Phi}\)).

sigma_l

Numeric matrix. Cholesky factorization (t(chol(sigma))) of the covariance matrix of volatility or randomness in the process (\(\boldsymbol{\Sigma}\)).

nu

Numeric vector. Vector of intercept values for the measurement model (\(\boldsymbol{\nu}\)).

lambda

Numeric matrix. Factor loading matrix linking the latent variables to the observed variables (\(\boldsymbol{\Lambda}\)).

theta_l

Numeric matrix. Cholesky factorization (t(chol(theta))) of the covariance matrix of the measurement error (\(\boldsymbol{\Theta}\)).

type

Integer. State space model type. See Details for more information.

x

List. Each element of the list is a matrix of covariates for each individual i in n. The number of columns in each matrix should be equal to time.

gamma

Numeric matrix. Matrix linking the covariates to the latent variables at current time point (\(\boldsymbol{\Gamma}\)).

kappa

Numeric matrix. Matrix linking the covariates to the observed variables at current time point (\(\boldsymbol{\kappa}\)).

mu0_fixed

Logical. If mu0_fixed = TRUE, fix the initial mean vector to mu0. If mu0_fixed = FALSE, mu0 is estimated.

sigma0_fixed

Logical. If sigma0_fixed = TRUE, fix the initial covariance matrix to tcrossprod(sigma0_l). If sigma0_fixed = FALSE, sigma0 is estimated.

alpha_level

Numeric vector. Significance level \(\alpha\).

optimization_flag

a flag (TRUE/FALSE) indicating whether optimization is to be done.

hessian_flag

a flag (TRUE/FALSE) indicating whether the Hessian matrix is to be calculated.

verbose

a flag (TRUE/FALSE) indicating whether more detailed intermediate output during the estimation process should be printed

weight_flag

a flag (TRUE/FALSE) indicating whether the negative log likelihood function should be weighted by the length of the time series for each individual

debug_flag

a flag (TRUE/FALSE) indicating whether users want additional dynr output that can be used for diagnostic purposes

perturb_flag

a flag (TRUE/FLASE) indicating whether to perturb the latent states during estimation. Only useful for ensemble forecasting.

xtol_rel

Stopping criteria option for parameter optimization. See dynr::dynr.model() for more details.

stopval

Stopping criteria option for parameter optimization. See dynr::dynr.model() for more details.

ftol_rel

Stopping criteria option for parameter optimization. See dynr::dynr.model() for more details.

ftol_abs

Stopping criteria option for parameter optimization. See dynr::dynr.model() for more details.

maxeval

Stopping criteria option for parameter optimization. See dynr::dynr.model() for more details.

maxtime

Stopping criteria option for parameter optimization. See dynr::dynr.model() for more details.

ncores

Positive integer. Number of cores to use. If ncores = NULL, use a single core. Consider using multiple cores when number of bootstrap samples R is a large value.

seed

Random seed.

clean

Logical. If clean = TRUE, delete intermediate files generated by the function.

Value

Returns an object of class bootstatespace which is a list with the following elements:

call

Function call.

args

Function arguments.

thetahatstar

Sampling distribution of \(\boldsymbol{\hat{\theta}}\).

vcov

Sampling variance-covariance matrix of \(\boldsymbol{\hat{\theta}}\).

est

Vector of estimated \(\boldsymbol{\hat{\theta}}\).

fun

Function used ("PBSSMLinSDEFixed").

method

Bootstrap method used ("parametric").

Details

Type 0

The measurement model is given by $$ \mathbf{y}_{i, t} = \boldsymbol{\nu} + \boldsymbol{\Lambda} \boldsymbol{\eta}_{i, t} + \boldsymbol{\varepsilon}_{i, t}, \quad \mathrm{with} \quad \boldsymbol{\varepsilon}_{i, t} \sim \mathcal{N} \left( \mathbf{0}, \boldsymbol{\Theta} \right) $$ where \(\mathbf{y}_{i, t}\), \(\boldsymbol{\eta}_{i, t}\), and \(\boldsymbol{\varepsilon}_{i, t}\) are random variables and \(\boldsymbol{\nu}\), \(\boldsymbol{\Lambda}\), and \(\boldsymbol{\Theta}\) are model parameters. \(\mathbf{y}_{i, t}\) represents a vector of observed random variables, \(\boldsymbol{\eta}_{i, t}\) a vector of latent random variables, and \(\boldsymbol{\varepsilon}_{i, t}\) a vector of random measurement errors, at time \(t\) and individual \(i\). \(\boldsymbol{\nu}\) denotes a vector of intercepts, \(\boldsymbol{\Lambda}\) a matrix of factor loadings, and \(\boldsymbol{\Theta}\) the covariance matrix of \(\boldsymbol{\varepsilon}\).

An alternative representation of the measurement error is given by $$ \boldsymbol{\varepsilon}_{i, t} = \boldsymbol{\Theta}^{\frac{1}{2}} \mathbf{z}_{i, t}, \quad \mathrm{with} \quad \mathbf{z}_{i, t} \sim \mathcal{N} \left( \mathbf{0}, \mathbf{I} \right) $$ where \(\mathbf{z}_{i, t}\) is a vector of independent standard normal random variables and \( \left( \boldsymbol{\Theta}^{\frac{1}{2}} \right) \left( \boldsymbol{\Theta}^{\frac{1}{2}} \right)^{\prime} = \boldsymbol{\Theta} . \)

The dynamic structure is given by $$ \mathrm{d} \boldsymbol{\eta}_{i, t} = \left( \boldsymbol{\iota} + \boldsymbol{\Phi} \boldsymbol{\eta}_{i, t} \right) \mathrm{d}t + \boldsymbol{\Sigma}^{\frac{1}{2}} \mathrm{d} \mathbf{W}_{i, t} $$ where \(\boldsymbol{\iota}\) is a term which is unobserved and constant over time, \(\boldsymbol{\Phi}\) is the drift matrix which represents the rate of change of the solution in the absence of any random fluctuations, \(\boldsymbol{\Sigma}\) is the matrix of volatility or randomness in the process, and \(\mathrm{d}\boldsymbol{W}\) is a Wiener process or Brownian motion, which represents random fluctuations.

Type 1

The measurement model is given by $$ \mathbf{y}_{i, t} = \boldsymbol{\nu} + \boldsymbol{\Lambda} \boldsymbol{\eta}_{i, t} + \boldsymbol{\varepsilon}_{i, t}, \quad \mathrm{with} \quad \boldsymbol{\varepsilon}_{i, t} \sim \mathcal{N} \left( \mathbf{0}, \boldsymbol{\Theta} \right) . $$

The dynamic structure is given by $$ \mathrm{d} \boldsymbol{\eta}_{i, t} = \left( \boldsymbol{\iota} + \boldsymbol{\Phi} \boldsymbol{\eta}_{i, t} \right) \mathrm{d}t + \boldsymbol{\Gamma} \mathbf{x}_{i, t} + \boldsymbol{\Sigma}^{\frac{1}{2}} \mathrm{d} \mathbf{W}_{i, t} $$ where \(\mathbf{x}_{i, t}\) represents a vector of covariates at time \(t\) and individual \(i\), and \(\boldsymbol{\Gamma}\) the coefficient matrix linking the covariates to the latent variables.

Type 2

The measurement model is given by $$ \mathbf{y}_{i, t} = \boldsymbol{\nu} + \boldsymbol{\Lambda} \boldsymbol{\eta}_{i, t} + \boldsymbol{\kappa} \mathbf{x}_{i, t} + \boldsymbol{\varepsilon}_{i, t}, \quad \mathrm{with} \quad \boldsymbol{\varepsilon}_{i, t} \sim \mathcal{N} \left( \mathbf{0}, \boldsymbol{\Theta} \right) $$ where \(\boldsymbol{\kappa}\) represents the coefficient matrix linking the covariates to the observed variables.

The dynamic structure is given by $$ \mathrm{d} \boldsymbol{\eta}_{i, t} = \left( \boldsymbol{\iota} + \boldsymbol{\Phi} \boldsymbol{\eta}_{i, t} \right) \mathrm{d}t + \boldsymbol{\Gamma} \mathbf{x}_{i, t} + \boldsymbol{\Sigma}^{\frac{1}{2}} \mathrm{d} \mathbf{W}_{i, t} . $$

State Space Parameterization

The state space parameters as a function of the linear stochastic differential equation model parameters are given by $$ \boldsymbol{\beta}_{\Delta t_{{l_{i}}}} = \exp{ \left( \Delta t \boldsymbol{\Phi} \right) } $$

$$ \boldsymbol{\alpha}_{\Delta t_{{l_{i}}}} = \boldsymbol{\Phi}^{-1} \left( \boldsymbol{\beta} - \mathbf{I}_{p} \right) \boldsymbol{\iota} $$

$$ \mathrm{vec} \left( \boldsymbol{\Psi}_{\Delta t_{{l_{i}}}} \right) = \left[ \left( \boldsymbol{\Phi} \otimes \mathbf{I}_{p} \right) + \left( \mathbf{I}_{p} \otimes \boldsymbol{\Phi} \right) \right] \left[ \exp \left( \left[ \left( \boldsymbol{\Phi} \otimes \mathbf{I}_{p} \right) + \left( \mathbf{I}_{p} \otimes \boldsymbol{\Phi} \right) \right] \Delta t \right) - \mathbf{I}_{p \times p} \right] \mathrm{vec} \left( \boldsymbol{\Sigma} \right) $$ where \(p\) is the number of latent variables and \(\Delta t\) is the time interval.

References

Chow, S.-M., Ho, M. R., Hamaker, E. L., & Dolan, C. V. (2010). Equivalence and differences between structural equation modeling and state-space modeling techniques. Structural Equation Modeling: A Multidisciplinary Journal, 17(2), 303–332. doi:10.1080/10705511003661553

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

# \donttest{
# prepare parameters
## number of individuals
n <- 5
## time points
time <- 50
delta_t <- 0.10
## dynamic structure
p <- 2
mu0 <- c(-3.0, 1.5)
sigma0 <- 0.001 * diag(p)
sigma0_l <- t(chol(sigma0))
iota <- c(0.317, 0.230)
phi <- matrix(
  data = c(
    -0.10,
    0.05,
    0.05,
    -0.10
  ),
  nrow = p
)
sigma <- matrix(
  data = c(
    2.79,
    0.06,
    0.06,
    3.27
  ),
  nrow = p
)
sigma_l <- t(chol(sigma))
## measurement model
k <- 2
nu <- rep(x = 0, times = k)
lambda <- diag(k)
theta <- 0.001 * diag(k)
theta_l <- t(chol(theta))

path <- tempdir()

pb <- PBSSMLinSDEFixed(
  R = 10L, # use at least 1000 in actual research
  path = path,
  prefix = "lse",
  n = n,
  time = time,
  delta_t = delta_t,
  mu0 = mu0,
  sigma0_l = sigma0_l,
  iota = iota,
  phi = phi,
  sigma_l = sigma_l,
  nu = nu,
  lambda = lambda,
  theta_l = theta_l,
  type = 0,
  ncores = 1, # consider using multiple cores
  seed = 42
)
print(pb)
#> Call:
#> PBSSMLinSDEFixed(R = 10L, path = path, prefix = "lse", n = n, 
#>     time = time, delta_t = delta_t, mu0 = mu0, sigma0_l = sigma0_l, 
#>     iota = iota, phi = phi, sigma_l = sigma_l, nu = nu, lambda = lambda, 
#>     theta_l = theta_l, type = 0, ncores = 1, seed = 42)
#> 
#> Parametric bootstrap confidence intervals.
#> type = "pc"
#>               est     se  R    2.5%   97.5%
#> phi_1_1    -0.100 0.1246 10 -0.4077 -0.0207
#> phi_2_1     0.050 0.2169 10 -0.1814  0.3589
#> phi_1_2     0.050 0.2563 10 -0.1668  0.5459
#> phi_2_2    -0.100 0.2802 10 -0.7522  0.0739
#> iota_1_1    0.317 0.7734 10 -1.4504  0.8892
#> iota_2_1    0.230 0.9178 10 -0.7594  1.8707
#> sigma_1_1   2.790 0.2025 10  2.4148  2.9673
#> sigma_2_1   0.060 0.1534 10 -0.2464  0.2022
#> sigma_2_2   3.270 0.1902 10  2.9992  3.4952
#> theta_1_1   0.001 0.0008 10  0.0000  0.0022
#> theta_2_2   0.001 0.0006 10  0.0000  0.0013
#> mu0_1_1    -3.000 0.0110 10 -3.0157 -2.9836
#> mu0_2_1     1.500 0.0250 10  1.4730  1.5375
#> sigma0_1_1  0.001 0.0009 10  0.0000  0.0021
#> sigma0_2_2  0.001 0.0020 10  0.0000  0.0047
summary(pb)
#> Call:
#> PBSSMLinSDEFixed(R = 10L, path = path, prefix = "lse", n = n, 
#>     time = time, delta_t = delta_t, mu0 = mu0, sigma0_l = sigma0_l, 
#>     iota = iota, phi = phi, sigma_l = sigma_l, nu = nu, lambda = lambda, 
#>     theta_l = theta_l, type = 0, ncores = 1, seed = 42)
#>               est     se  R    2.5%   97.5%
#> phi_1_1    -0.100 0.1246 10 -0.4077 -0.0207
#> phi_2_1     0.050 0.2169 10 -0.1814  0.3589
#> phi_1_2     0.050 0.2563 10 -0.1668  0.5459
#> phi_2_2    -0.100 0.2802 10 -0.7522  0.0739
#> iota_1_1    0.317 0.7734 10 -1.4504  0.8892
#> iota_2_1    0.230 0.9178 10 -0.7594  1.8707
#> sigma_1_1   2.790 0.2025 10  2.4148  2.9673
#> sigma_2_1   0.060 0.1534 10 -0.2464  0.2022
#> sigma_2_2   3.270 0.1902 10  2.9992  3.4952
#> theta_1_1   0.001 0.0008 10  0.0000  0.0022
#> theta_2_2   0.001 0.0006 10  0.0000  0.0013
#> mu0_1_1    -3.000 0.0110 10 -3.0157 -2.9836
#> mu0_2_1     1.500 0.0250 10  1.4730  1.5375
#> sigma0_1_1  0.001 0.0009 10  0.0000  0.0021
#> sigma0_2_2  0.001 0.0020 10  0.0000  0.0047
confint(pb)
#>                    2.5 %       97.5 %
#> phi_1_1    -4.077282e-01 -0.020726350
#> phi_2_1    -1.813916e-01  0.358896477
#> phi_1_2    -1.668021e-01  0.545918950
#> phi_2_2    -7.522355e-01  0.073878420
#> iota_1_1   -1.450424e+00  0.889158714
#> iota_2_1   -7.593950e-01  1.870666652
#> sigma_1_1   2.414810e+00  2.967317406
#> sigma_2_1  -2.464483e-01  0.202162126
#> sigma_2_2   2.999248e+00  3.495216809
#> theta_1_1   1.017475e-11  0.002190168
#> theta_2_2   2.439758e-14  0.001348227
#> mu0_1_1    -3.015689e+00 -2.983648305
#> mu0_2_1     1.473001e+00  1.537465316
#> sigma0_1_1  5.516864e-28  0.002105120
#> sigma0_2_2  2.465699e-20  0.004749849
vcov(pb)
#>                  phi_1_1       phi_2_1       phi_1_2       phi_2_2
#> phi_1_1     1.552256e-02 -1.214203e-02 -2.337715e-02 -1.983527e-03
#> phi_2_1    -1.214203e-02  4.703328e-02  3.433472e-02 -3.851923e-02
#> phi_1_2    -2.337715e-02  3.433472e-02  6.571480e-02 -1.225741e-02
#> phi_2_2    -1.983527e-03 -3.851923e-02 -1.225741e-02  7.849204e-02
#> iota_1_1    7.504775e-02 -7.361112e-02 -1.655828e-01 -3.646284e-02
#> iota_2_1   -4.276972e-02  1.253285e-01  1.086847e-01 -2.098571e-01
#> sigma_1_1   1.037976e-02  1.473525e-02 -2.071599e-02 -4.302432e-02
#> sigma_2_1   1.237386e-02 -6.654653e-03 -1.146227e-02 -2.044285e-02
#> sigma_2_2  -4.634435e-03  1.900901e-02  2.181110e-02 -5.262894e-03
#> theta_1_1  -2.367485e-05  4.962387e-05  2.166866e-05  1.206335e-05
#> theta_2_2   2.420195e-05 -1.989012e-05 -1.252671e-05  6.248239e-05
#> mu0_1_1     6.592888e-04 -2.549038e-04 -1.338046e-03 -1.319097e-03
#> mu0_2_1     8.431330e-04  3.906179e-04 -1.703047e-03  1.831875e-03
#> sigma0_1_1  4.025716e-05 -1.714202e-05 -6.940889e-05 -1.252895e-04
#> sigma0_2_2 -5.278016e-05  1.374084e-04  1.433501e-06 -2.316630e-04
#>                 iota_1_1      iota_2_1     sigma_1_1     sigma_2_1
#> phi_1_1     7.504775e-02 -0.0427697233  1.037976e-02  1.237386e-02
#> phi_2_1    -7.361112e-02  0.1253285161  1.473525e-02 -6.654653e-03
#> phi_1_2    -1.655828e-01  0.1086846549 -2.071599e-02 -1.146227e-02
#> phi_2_2    -3.646284e-02 -0.2098570504 -4.302432e-02 -2.044285e-02
#> iota_1_1    5.981176e-01 -0.0539415490  8.032409e-02  7.237663e-02
#> iota_2_1   -5.394155e-02  0.8424087126  7.522060e-02  4.145535e-02
#> sigma_1_1   8.032409e-02  0.0752205996  4.100248e-02  1.445459e-02
#> sigma_2_1   7.237663e-02  0.0414553460  1.445459e-02  2.352967e-02
#> sigma_2_2  -8.864587e-02 -0.0101400481  3.760631e-03 -6.614318e-03
#> theta_1_1  -1.359188e-04 -0.0001826804 -1.877780e-07 -7.964207e-05
#> theta_2_2  -2.303878e-05 -0.0002331378 -1.370408e-05 -2.538503e-06
#> mu0_1_1     4.343448e-03  0.0013614155  1.415151e-03  9.458768e-04
#> mu0_2_1     2.259898e-03 -0.0112586849  6.469714e-04 -5.023952e-04
#> sigma0_1_1  3.146641e-04  0.0003790998  1.002904e-04  1.005176e-04
#> sigma0_2_2  1.221559e-04  0.0009135827  1.544907e-04  3.352876e-05
#>                sigma_2_2     theta_1_1     theta_2_2       mu0_1_1
#> phi_1_1    -4.634435e-03 -2.367485e-05  2.420195e-05  6.592888e-04
#> phi_2_1     1.900901e-02  4.962387e-05 -1.989012e-05 -2.549038e-04
#> phi_1_2     2.181110e-02  2.166866e-05 -1.252671e-05 -1.338046e-03
#> phi_2_2    -5.262894e-03  1.206335e-05  6.248239e-05 -1.319097e-03
#> iota_1_1   -8.864587e-02 -1.359188e-04 -2.303878e-05  4.343448e-03
#> iota_2_1   -1.014005e-02 -1.826804e-04 -2.331378e-04  1.361415e-03
#> sigma_1_1   3.760631e-03 -1.877780e-07 -1.370408e-05  1.415151e-03
#> sigma_2_1  -6.614318e-03 -7.964207e-05 -2.538503e-06  9.458768e-04
#> sigma_2_2   3.615920e-02  4.927267e-05  2.562057e-05  1.388412e-04
#> theta_1_1   4.927267e-05  7.174432e-07 -3.831374e-08 -2.759259e-06
#> theta_2_2   2.562057e-05 -3.831374e-08  3.732333e-07 -2.948589e-06
#> mu0_1_1     1.388412e-04 -2.759259e-06 -2.948589e-06  1.208080e-04
#> mu0_2_1     9.985030e-04  7.573211e-06  7.671619e-06 -6.663468e-05
#> sigma0_1_1 -2.961326e-05 -5.936253e-07 -1.245753e-07  7.684620e-06
#> sigma0_2_2  1.738283e-05 -4.680011e-07 -8.230699e-07  1.483931e-05
#>                  mu0_2_1    sigma0_1_1    sigma0_2_2
#> phi_1_1     8.431330e-04  4.025716e-05 -5.278016e-05
#> phi_2_1     3.906179e-04 -1.714202e-05  1.374084e-04
#> phi_1_2    -1.703047e-03 -6.940889e-05  1.433501e-06
#> phi_2_2     1.831875e-03 -1.252895e-04 -2.316630e-04
#> iota_1_1    2.259898e-03  3.146641e-04  1.221559e-04
#> iota_2_1   -1.125868e-02  3.790998e-04  9.135827e-04
#> sigma_1_1   6.469714e-04  1.002904e-04  1.544907e-04
#> sigma_2_1  -5.023952e-04  1.005176e-04  3.352876e-05
#> sigma_2_2   9.985030e-04 -2.961326e-05  1.738283e-05
#> theta_1_1   7.573211e-06 -5.936253e-07 -4.680011e-07
#> theta_2_2   7.671619e-06 -1.245753e-07 -8.230699e-07
#> mu0_1_1    -6.663468e-05  7.684620e-06  1.483931e-05
#> mu0_2_1     6.230784e-04 -1.079420e-05 -2.047741e-05
#> sigma0_1_1 -1.079420e-05  8.397469e-07  1.133023e-06
#> sigma0_2_2 -2.047741e-05  1.133023e-06  4.043116e-06
coef(pb)
#>    phi_1_1    phi_2_1    phi_1_2    phi_2_2   iota_1_1   iota_2_1  sigma_1_1 
#>     -0.100      0.050      0.050     -0.100      0.317      0.230      2.790 
#>  sigma_2_1  sigma_2_2  theta_1_1  theta_2_2    mu0_1_1    mu0_2_1 sigma0_1_1 
#>      0.060      3.270      0.001      0.001     -3.000      1.500      0.001 
#> sigma0_2_2 
#>      0.001 
print(pb, type = "bc") # bias-corrected
#> Call:
#> PBSSMLinSDEFixed(R = 10L, path = path, prefix = "lse", n = n, 
#>     time = time, delta_t = delta_t, mu0 = mu0, sigma0_l = sigma0_l, 
#>     iota = iota, phi = phi, sigma_l = sigma_l, nu = nu, lambda = lambda, 
#>     theta_l = theta_l, type = 0, ncores = 1, seed = 42)
#> 
#> Parametric bootstrap confidence intervals.
#> type = "bc"
#>               est     se  R    2.5%   97.5%
#> phi_1_1    -0.100 0.1246 10 -0.2823  0.0009
#> phi_2_1     0.050 0.2169 10 -0.1814  0.3589
#> phi_1_2     0.050 0.2563 10 -0.1716  0.4011
#> phi_2_2    -0.100 0.2802 10 -0.3240  0.0824
#> iota_1_1    0.317 0.7734 10 -0.8096  0.9456
#> iota_2_1    0.230 0.9178 10 -0.8578  1.5604
#> sigma_1_1   2.790 0.2025 10  2.4148  2.9673
#> sigma_2_1   0.060 0.1534 10 -0.1441  0.2385
#> sigma_2_2   3.270 0.1902 10  2.9992  3.4952
#> theta_1_1   0.001 0.0008 10  0.0000  0.0021
#> theta_2_2   0.001 0.0006 10  0.0000  0.0014
#> mu0_1_1    -3.000 0.0110 10 -3.0157 -2.9862
#> mu0_2_1     1.500 0.0250 10  1.4723  1.5333
#> sigma0_1_1  0.001 0.0009 10  0.0000  0.0021
#> sigma0_2_2  0.001 0.0020 10  0.0000  0.0048
summary(pb, type = "bc")
#> Call:
#> PBSSMLinSDEFixed(R = 10L, path = path, prefix = "lse", n = n, 
#>     time = time, delta_t = delta_t, mu0 = mu0, sigma0_l = sigma0_l, 
#>     iota = iota, phi = phi, sigma_l = sigma_l, nu = nu, lambda = lambda, 
#>     theta_l = theta_l, type = 0, ncores = 1, seed = 42)
#>               est     se  R    2.5%   97.5%
#> phi_1_1    -0.100 0.1246 10 -0.2823  0.0009
#> phi_2_1     0.050 0.2169 10 -0.1814  0.3589
#> phi_1_2     0.050 0.2563 10 -0.1716  0.4011
#> phi_2_2    -0.100 0.2802 10 -0.3240  0.0824
#> iota_1_1    0.317 0.7734 10 -0.8096  0.9456
#> iota_2_1    0.230 0.9178 10 -0.8578  1.5604
#> sigma_1_1   2.790 0.2025 10  2.4148  2.9673
#> sigma_2_1   0.060 0.1534 10 -0.1441  0.2385
#> sigma_2_2   3.270 0.1902 10  2.9992  3.4952
#> theta_1_1   0.001 0.0008 10  0.0000  0.0021
#> theta_2_2   0.001 0.0006 10  0.0000  0.0014
#> mu0_1_1    -3.000 0.0110 10 -3.0157 -2.9862
#> mu0_2_1     1.500 0.0250 10  1.4723  1.5333
#> sigma0_1_1  0.001 0.0009 10  0.0000  0.0021
#> sigma0_2_2  0.001 0.0020 10  0.0000  0.0048
confint(pb, type = "bc")
#>                    2.5 %        97.5 %
#> phi_1_1    -2.823085e-01  0.0008998844
#> phi_2_1    -1.813916e-01  0.3588964768
#> phi_1_2    -1.715504e-01  0.4011433985
#> phi_2_2    -3.239992e-01  0.0823962255
#> iota_1_1   -8.095672e-01  0.9456081688
#> iota_2_1   -8.577605e-01  1.5603551714
#> sigma_1_1   2.414810e+00  2.9673174061
#> sigma_2_1  -1.440736e-01  0.2384939164
#> sigma_2_2   2.999248e+00  3.4952168086
#> theta_1_1   2.777357e-12  0.0020779361
#> theta_2_2   4.697258e-12  0.0013545688
#> mu0_1_1    -3.015722e+00 -2.9862201382
#> mu0_2_1     1.472278e+00  1.5332783257
#> sigma0_1_1  2.129691e-24  0.0021456288
#> sigma0_2_2  3.763664e-16  0.0048338734
# }