Parametric Bootstrap for the Vector Autoregressive Model (Fixed Parameters)

This function simulates data from a vector autoregressive model using a state-space model parameterization and fits the model using the dynr package. The process is repeated R times. It assumes that the parameters remain constant across individuals and over time. At the moment, the function only supports type = 0.

Usage

PBSSMVARFixed(
  R,
  path,
  prefix,
  n,
  time,
  mu0,
  sigma0_l,
  alpha,
  beta,
  psi_l,
  type = 0,
  x = NULL,
  gamma = NULL,
  mu0_fixed = FALSE,
  sigma0_fixed = FALSE,
  alpha_level = 0.05,
  optimization_flag = TRUE,
  hessian_flag = FALSE,
  verbose = FALSE,
  weight_flag = FALSE,
  debug_flag = FALSE,
  perturb_flag = FALSE,
  xtol_rel = 1e-07,
  stopval = -9999,
  ftol_rel = -1,
  ftol_abs = -1,
  maxeval = as.integer(-1),
  maxtime = -1,
  ncores = NULL,
  seed = NULL,
  clean = TRUE
)

Arguments

R

Positive integer. Number of bootstrap samples.

path

Path to a directory to store bootstrap samples and estimates.

prefix

Character string. Prefix used for the file names for the bootstrap samples and estimates.

n

Positive integer. Number of individuals.

time

Positive integer. Number of time points.

mu0

Numeric vector. Mean of initial latent variable values (\(\boldsymbol{\mu}_{\boldsymbol{\eta} \mid 0}\)).

sigma0_l

Numeric matrix. Cholesky factorization (t(chol(sigma0))) of the covariance matrix of initial latent variable values (\(\boldsymbol{\Sigma}_{\boldsymbol{\eta} \mid 0}\)).

alpha

Numeric vector. Vector of constant values for the dynamic model (\(\boldsymbol{\alpha}\)).

beta

Numeric matrix. Transition matrix relating the values of the latent variables at the previous to the current time point (\(\boldsymbol{\beta}\)).

psi_l

Numeric matrix. Cholesky factorization (t(chol(psi))) of the covariance matrix of the process noise (\(\boldsymbol{\Psi}\)).

type

Integer. State space model type. See Details for more information.

x

List. Each element of the list is a matrix of covariates for each individual i in n. The number of columns in each matrix should be equal to time.

gamma

Numeric matrix. Matrix linking the covariates to the latent variables at current time point (\(\boldsymbol{\Gamma}\)).

mu0_fixed

Logical. If mu0_fixed = TRUE, fix the initial mean vector to mu0. If mu0_fixed = FALSE, mu0 is estimated.

sigma0_fixed

Logical. If sigma0_fixed = TRUE, fix the initial covariance matrix to tcrossprod(sigma0_l). If sigma0_fixed = FALSE, sigma0 is estimated.

alpha_level

Numeric vector. Significance level \(\alpha\).

optimization_flag

a flag (TRUE/FALSE) indicating whether optimization is to be done.

hessian_flag

a flag (TRUE/FALSE) indicating whether the Hessian matrix is to be calculated.

verbose

a flag (TRUE/FALSE) indicating whether more detailed intermediate output during the estimation process should be printed

weight_flag

a flag (TRUE/FALSE) indicating whether the negative log likelihood function should be weighted by the length of the time series for each individual

debug_flag

a flag (TRUE/FALSE) indicating whether users want additional dynr output that can be used for diagnostic purposes

perturb_flag

a flag (TRUE/FLASE) indicating whether to perturb the latent states during estimation. Only useful for ensemble forecasting.

xtol_rel

Stopping criteria option for parameter optimization. See dynr::dynr.model() for more details.

stopval

Stopping criteria option for parameter optimization. See dynr::dynr.model() for more details.

ftol_rel

Stopping criteria option for parameter optimization. See dynr::dynr.model() for more details.

ftol_abs

Stopping criteria option for parameter optimization. See dynr::dynr.model() for more details.

maxeval

Stopping criteria option for parameter optimization. See dynr::dynr.model() for more details.

maxtime

Stopping criteria option for parameter optimization. See dynr::dynr.model() for more details.

ncores

Positive integer. Number of cores to use. If ncores = NULL, use a single core. Consider using multiple cores when number of bootstrap samples R is a large value.

seed

Random seed.

clean

Logical. If clean = TRUE, delete intermediate files generated by the function.

Value

Returns an object of class bootstatespace which is a list with the following elements:

call

Function call.

args

Function arguments.

thetahatstar

Sampling distribution of \(\boldsymbol{\hat{\theta}}\).

vcov

Sampling variance-covariance matrix of \(\boldsymbol{\hat{\theta}}\).

est

Vector of estimated \(\boldsymbol{\hat{\theta}}\).

fun

Function used ("PBSSMVARFixed").

method

Bootstrap method used ("parametric").

Details

Type 0

The measurement model is given by $$ \mathbf{y}_{i, t} = \boldsymbol{\eta}_{i, t} $$ where \(\mathbf{y}_{i, t}\) represents a vector of observed variables and \(\boldsymbol{\eta}_{i, t}\) a vector of latent variables for individual \(i\) and time \(t\). Since the observed and latent variables are equal, we only generate data from the dynamic structure.

The dynamic structure is given by $$ \boldsymbol{\eta}_{i, t} = \boldsymbol{\alpha} + \boldsymbol{\beta} \boldsymbol{\eta}_{i, t - 1} + \boldsymbol{\zeta}_{i, t}, \quad \mathrm{with} \quad \boldsymbol{\zeta}_{i, t} \sim \mathcal{N} \left( \mathbf{0}, \boldsymbol{\Psi} \right) $$ where \(\boldsymbol{\eta}_{i, t}\), \(\boldsymbol{\eta}_{i, t - 1}\), and \(\boldsymbol{\zeta}_{i, t}\) are random variables, and \(\boldsymbol{\alpha}\), \(\boldsymbol{\beta}\), and \(\boldsymbol{\Psi}\) are model parameters. Here, \(\boldsymbol{\eta}_{i, t}\) is a vector of latent variables at time \(t\) and individual \(i\), \(\boldsymbol{\eta}_{i, t - 1}\) represents a vector of latent variables at time \(t - 1\) and individual \(i\), and \(\boldsymbol{\zeta}_{i, t}\) represents a vector of dynamic noise at time \(t\) and individual \(i\). \(\boldsymbol{\alpha}\) denotes a vector of intercepts, \(\boldsymbol{\beta}\) a matrix of autoregression and cross regression coefficients, and \(\boldsymbol{\Psi}\) the covariance matrix of \(\boldsymbol{\zeta}_{i, t}\).

An alternative representation of the dynamic noise is given by $$ \boldsymbol{\zeta}_{i, t} = \boldsymbol{\Psi}^{\frac{1}{2}} \mathbf{z}_{i, t}, \quad \mathrm{with} \quad \mathbf{z}_{i, t} \sim \mathcal{N} \left( \mathbf{0}, \mathbf{I} \right) $$ where \( \left( \boldsymbol{\Psi}^{\frac{1}{2}} \right) \left( \boldsymbol{\Psi}^{\frac{1}{2}} \right)^{\prime} = \boldsymbol{\Psi} . \)

Type 1

The measurement model is given by $$ \mathbf{y}_{i, t} = \boldsymbol{\eta}_{i, t} . $$

The dynamic structure is given by $$ \boldsymbol{\eta}_{i, t} = \boldsymbol{\alpha} + \boldsymbol{\beta} \boldsymbol{\eta}_{i, t - 1} + \boldsymbol{\Gamma} \mathbf{x}_{i, t} + \boldsymbol{\zeta}_{i, t}, \quad \mathrm{with} \quad \boldsymbol{\zeta}_{i, t} \sim \mathcal{N} \left( \mathbf{0}, \boldsymbol{\Psi} \right) $$ where \(\mathbf{x}_{i, t}\) represents a vector of covariates at time \(t\) and individual \(i\), and \(\boldsymbol{\Gamma}\) the coefficient matrix linking the covariates to the latent variables.

References

Chow, S.-M., Ho, M. R., Hamaker, E. L., & Dolan, C. V. (2010). Equivalence and differences between structural equation modeling and state-space modeling techniques. Structural Equation Modeling: A Multidisciplinary Journal, 17(2), 303–332. doi:10.1080/10705511003661553

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

# \donttest{
# prepare parameters
## number of individuals
n <- 5
## time points
time <- 50
## dynamic structure
p <- 3
mu0 <- rep(x = 0, times = p)
sigma0 <- 0.001 * diag(p)
sigma0_l <- t(chol(sigma0))
alpha <- rep(x = 0, times = p)
beta <- 0.50 * diag(p)
psi <- 0.001 * diag(p)
psi_l <- t(chol(psi))

path <- tempdir()

pb <- PBSSMVARFixed(
  R = 10L, # use at least 1000 in actual research
  path = path,
  prefix = "var",
  n = n,
  time = time,
  mu0 = mu0,
  sigma0_l = sigma0_l,
  alpha = alpha,
  beta = beta,
  psi_l = psi_l,
  type = 0,
  ncores = 1, # consider using multiple cores
  seed = 42
)
print(pb)
#> Call:
#> PBSSMVARFixed(R = 10L, path = path, prefix = "var", n = n, time = time, 
#>     mu0 = mu0, sigma0_l = sigma0_l, alpha = alpha, beta = beta, 
#>     psi_l = psi_l, type = 0, ncores = 1, seed = 42)
#> 
#> Parametric bootstrap confidence intervals.
#> type = "pc"
#>              est     se  R    2.5%  97.5%
#> beta_1_1   0.500 0.0479 10  0.4116 0.5553
#> beta_2_1   0.000 0.0430 10 -0.0404 0.0828
#> beta_3_1   0.000 0.0719 10 -0.0782 0.1313
#> beta_1_2   0.000 0.0697 10 -0.0681 0.1182
#> beta_2_2   0.500 0.0546 10  0.4440 0.6032
#> beta_3_2   0.000 0.0371 10 -0.0102 0.0967
#> beta_1_3   0.000 0.0450 10 -0.0725 0.0522
#> beta_2_3   0.000 0.0411 10 -0.0574 0.0705
#> beta_3_3   0.500 0.0824 10  0.3106 0.5658
#> psi_1_1    0.001 0.0001 10  0.0009 0.0011
#> psi_2_2    0.001 0.0001 10  0.0009 0.0011
#> psi_3_3    0.001 0.0001 10  0.0009 0.0011
#> mu0_1_1    0.000 0.0083 10 -0.0094 0.0114
#> mu0_2_1    0.000 0.0103 10 -0.0108 0.0207
#> mu0_3_1    0.000 0.0132 10 -0.0246 0.0155
#> sigma0_1_1 0.001 0.0005 10  0.0002 0.0018
#> sigma0_2_2 0.001 0.0012 10  0.0004 0.0036
#> sigma0_3_3 0.001 0.0006 10  0.0005 0.0022
summary(pb)
#>              est     se  R    2.5%  97.5%
#> beta_1_1   0.500 0.0479 10  0.4116 0.5553
#> beta_2_1   0.000 0.0430 10 -0.0404 0.0828
#> beta_3_1   0.000 0.0719 10 -0.0782 0.1313
#> beta_1_2   0.000 0.0697 10 -0.0681 0.1182
#> beta_2_2   0.500 0.0546 10  0.4440 0.6032
#> beta_3_2   0.000 0.0371 10 -0.0102 0.0967
#> beta_1_3   0.000 0.0450 10 -0.0725 0.0522
#> beta_2_3   0.000 0.0411 10 -0.0574 0.0705
#> beta_3_3   0.500 0.0824 10  0.3106 0.5658
#> psi_1_1    0.001 0.0001 10  0.0009 0.0011
#> psi_2_2    0.001 0.0001 10  0.0009 0.0011
#> psi_3_3    0.001 0.0001 10  0.0009 0.0011
#> mu0_1_1    0.000 0.0083 10 -0.0094 0.0114
#> mu0_2_1    0.000 0.0103 10 -0.0108 0.0207
#> mu0_3_1    0.000 0.0132 10 -0.0246 0.0155
#> sigma0_1_1 0.001 0.0005 10  0.0002 0.0018
#> sigma0_2_2 0.001 0.0012 10  0.0004 0.0036
#> sigma0_3_3 0.001 0.0006 10  0.0005 0.0022
confint(pb)
#>                    2.5 %      97.5 %
#> beta_1_1    0.4116181715 0.555290574
#> beta_2_1   -0.0403928322 0.082848108
#> beta_3_1   -0.0781574901 0.131326844
#> beta_1_2   -0.0681340006 0.118237420
#> beta_2_2    0.4439997593 0.603233016
#> beta_3_2   -0.0101631805 0.096683016
#> beta_1_3   -0.0724607669 0.052155419
#> beta_2_3   -0.0574483499 0.070495533
#> beta_3_3    0.3106499503 0.565785361
#> psi_1_1     0.0008514919 0.001069493
#> psi_2_2     0.0008984977 0.001061206
#> psi_3_3     0.0008957730 0.001084173
#> mu0_1_1    -0.0093538510 0.011387247
#> mu0_2_1    -0.0107924456 0.020685178
#> mu0_3_1    -0.0246156468 0.015541177
#> sigma0_1_1  0.0002063937 0.001753763
#> sigma0_2_2  0.0003663549 0.003572860
#> sigma0_3_3  0.0005407295 0.002243706
vcov(pb)
#>                 beta_1_1      beta_2_1      beta_3_1      beta_1_2
#> beta_1_1    2.298086e-03 -5.946947e-04 -1.588033e-03 -8.622130e-04
#> beta_2_1   -5.946947e-04  1.852729e-03  2.650435e-03 -1.550321e-04
#> beta_3_1   -1.588033e-03  2.650435e-03  5.166260e-03 -1.285603e-04
#> beta_1_2   -8.622130e-04 -1.550321e-04 -1.285603e-04  4.855526e-03
#> beta_2_2    6.899260e-04  8.070575e-04  1.553413e-03  6.133108e-04
#> beta_3_2    3.367613e-04 -1.001946e-04  4.889234e-04 -2.457563e-04
#> beta_1_3    4.616224e-04 -2.840657e-04 -1.058703e-03  1.220404e-03
#> beta_2_3    3.707927e-04 -1.212402e-06 -5.186974e-04  1.376664e-03
#> beta_3_3   -7.594566e-04 -4.098391e-06 -1.524120e-04  9.880377e-04
#> psi_1_1    -1.661606e-06  7.127546e-07  8.318978e-07 -2.271265e-06
#> psi_2_2     5.244801e-07  1.247412e-07  2.238045e-07  1.264535e-06
#> psi_3_3    -1.856750e-06  4.940852e-07  1.127305e-06  1.216409e-06
#> mu0_1_1    -1.724581e-04 -9.382073e-05 -1.019350e-04  2.291777e-04
#> mu0_2_1    -1.694469e-04  1.204791e-04  3.112090e-04  3.575750e-04
#> mu0_3_1     1.377551e-04  1.724709e-04  1.730457e-04 -5.758640e-04
#> sigma0_1_1  2.866472e-06  5.582489e-06 -1.855611e-06 -2.311721e-05
#> sigma0_2_2 -1.861969e-05 -6.285895e-06 -7.878012e-06 -1.485582e-05
#> sigma0_3_3  5.638744e-06  5.116476e-06  1.224554e-05 -3.687752e-05
#>                 beta_2_2      beta_3_2      beta_1_3      beta_2_3
#> beta_1_1    6.899260e-04  3.367613e-04  4.616224e-04  3.707927e-04
#> beta_2_1    8.070575e-04 -1.001946e-04 -2.840657e-04 -1.212402e-06
#> beta_3_1    1.553413e-03  4.889234e-04 -1.058703e-03 -5.186974e-04
#> beta_1_2    6.133108e-04 -2.457563e-04  1.220404e-03  1.376664e-03
#> beta_2_2    2.978467e-03 -4.008152e-05  4.316511e-04  3.805633e-04
#> beta_3_2   -4.008152e-05  1.375060e-03 -7.350628e-04 -2.984134e-04
#> beta_1_3    4.316511e-04 -7.350628e-04  2.026186e-03  9.021806e-04
#> beta_2_3    3.805633e-04 -2.984134e-04  9.021806e-04  1.687769e-03
#> beta_3_3    1.083996e-03 -1.103972e-03 -1.155997e-03 -2.355697e-04
#> psi_1_1    -2.936293e-06 -2.866951e-07 -2.895984e-08 -9.442857e-07
#> psi_2_2     3.824225e-07  1.322813e-06  1.812758e-07 -1.401687e-07
#> psi_3_3    -1.491368e-06  2.872840e-07 -2.333341e-07  1.039667e-06
#> mu0_1_1    -1.218306e-04  4.038472e-06  1.057204e-04 -5.024007e-05
#> mu0_2_1    -2.892627e-05  2.068456e-04  1.494827e-06  9.319722e-05
#> mu0_3_1    -3.267230e-05  4.428562e-05  2.207153e-06  1.296002e-04
#> sigma0_1_1 -8.968996e-06 -1.221431e-06 -1.000590e-05 -6.728590e-06
#> sigma0_2_2  4.120067e-07 -7.511271e-06  8.447553e-06 -4.530228e-07
#> sigma0_3_3  7.080961e-06  2.171812e-06 -9.408207e-06 -1.789169e-05
#>                 beta_3_3       psi_1_1       psi_2_2       psi_3_3
#> beta_1_1   -7.594566e-04 -1.661606e-06  5.244801e-07 -1.856750e-06
#> beta_2_1   -4.098391e-06  7.127546e-07  1.247412e-07  4.940852e-07
#> beta_3_1   -1.524120e-04  8.318978e-07  2.238045e-07  1.127305e-06
#> beta_1_2    9.880377e-04 -2.271265e-06  1.264535e-06  1.216409e-06
#> beta_2_2    1.083996e-03 -2.936293e-06  3.824225e-07 -1.491368e-06
#> beta_3_2   -1.103972e-03 -2.866951e-07  1.322813e-06  2.872840e-07
#> beta_1_3   -1.155997e-03 -2.895984e-08  1.812758e-07 -2.333341e-07
#> beta_2_3   -2.355697e-04 -9.442857e-07 -1.401687e-07  1.039667e-06
#> beta_3_3    6.793019e-03 -3.271143e-06 -2.841936e-07 -1.591097e-07
#> psi_1_1    -3.271143e-06  6.257162e-09 -1.016943e-09  1.995431e-09
#> psi_2_2    -2.841936e-07 -1.016943e-09  3.356282e-09  4.658279e-10
#> psi_3_3    -1.591097e-07  1.995431e-09  4.658279e-10  4.432737e-09
#> mu0_1_1    -2.138337e-04  1.722138e-07  3.621344e-08  9.907652e-09
#> mu0_2_1    -4.321791e-04  1.110986e-07  2.320145e-07  3.138911e-07
#> mu0_3_1    -3.717613e-04  4.210829e-07 -8.960295e-08  2.770162e-07
#> sigma0_1_1 -4.795336e-07  1.562190e-08 -7.522821e-09 -8.363358e-09
#> sigma0_2_2 -4.810817e-06  2.357452e-08 -1.722137e-08  6.312961e-09
#> sigma0_3_3  8.164963e-07  9.412104e-09 -3.924849e-09 -1.837240e-08
#>                  mu0_1_1       mu0_2_1       mu0_3_1    sigma0_1_1
#> beta_1_1   -1.724581e-04 -1.694469e-04  1.377551e-04  2.866472e-06
#> beta_2_1   -9.382073e-05  1.204791e-04  1.724709e-04  5.582489e-06
#> beta_3_1   -1.019350e-04  3.112090e-04  1.730457e-04 -1.855611e-06
#> beta_1_2    2.291777e-04  3.575750e-04 -5.758640e-04 -2.311721e-05
#> beta_2_2   -1.218306e-04 -2.892627e-05 -3.267230e-05 -8.968996e-06
#> beta_3_2    4.038472e-06  2.068456e-04  4.428562e-05 -1.221431e-06
#> beta_1_3    1.057204e-04  1.494827e-06  2.207153e-06 -1.000590e-05
#> beta_2_3   -5.024007e-05  9.319722e-05  1.296002e-04 -6.728590e-06
#> beta_3_3   -2.138337e-04 -4.321791e-04 -3.717613e-04 -4.795336e-07
#> psi_1_1     1.722138e-07  1.110986e-07  4.210829e-07  1.562190e-08
#> psi_2_2     3.621344e-08  2.320145e-07 -8.960295e-08 -7.522821e-09
#> psi_3_3     9.907652e-09  3.138911e-07  2.770162e-07 -8.363358e-09
#> mu0_1_1     6.811139e-05  4.001262e-05 -5.846844e-05 -4.793451e-07
#> mu0_2_1     4.001262e-05  1.058619e-04 -1.670120e-05 -1.510791e-06
#> mu0_3_1    -5.846844e-05 -1.670120e-05  1.742157e-04  1.901602e-06
#> sigma0_1_1 -4.793451e-07 -1.510791e-06  1.901602e-06  2.763533e-07
#> sigma0_2_2  4.741424e-06 -4.438289e-07  2.363057e-06  9.901790e-08
#> sigma0_3_3 -2.030631e-06 -3.317229e-06  3.183708e-06  1.159188e-07
#>               sigma0_2_2    sigma0_3_3
#> beta_1_1   -1.861969e-05  5.638744e-06
#> beta_2_1   -6.285895e-06  5.116476e-06
#> beta_3_1   -7.878012e-06  1.224554e-05
#> beta_1_2   -1.485582e-05 -3.687752e-05
#> beta_2_2    4.120067e-07  7.080961e-06
#> beta_3_2   -7.511271e-06  2.171812e-06
#> beta_1_3    8.447553e-06 -9.408207e-06
#> beta_2_3   -4.530228e-07 -1.789169e-05
#> beta_3_3   -4.810817e-06  8.164963e-07
#> psi_1_1     2.357452e-08  9.412104e-09
#> psi_2_2    -1.722137e-08 -3.924849e-09
#> psi_3_3     6.312961e-09 -1.837240e-08
#> mu0_1_1     4.741424e-06 -2.030631e-06
#> mu0_2_1    -4.438289e-07 -3.317229e-06
#> mu0_3_1     2.363057e-06  3.183708e-06
#> sigma0_1_1  9.901790e-08  1.159188e-07
#> sigma0_2_2  1.336808e-06  6.526429e-08
#> sigma0_3_3  6.526429e-08  3.982781e-07
coef(pb)
#>   beta_1_1   beta_2_1   beta_3_1   beta_1_2   beta_2_2   beta_3_2   beta_1_3 
#>      0.500      0.000      0.000      0.000      0.500      0.000      0.000 
#>   beta_2_3   beta_3_3    psi_1_1    psi_2_2    psi_3_3    mu0_1_1    mu0_2_1 
#>      0.000      0.500      0.001      0.001      0.001      0.000      0.000 
#>    mu0_3_1 sigma0_1_1 sigma0_2_2 sigma0_3_3 
#>      0.000      0.001      0.001      0.001 
print(pb, type = "bc") # bias-corrected
#> Call:
#> PBSSMVARFixed(R = 10L, path = path, prefix = "var", n = n, time = time, 
#>     mu0 = mu0, sigma0_l = sigma0_l, alpha = alpha, beta = beta, 
#>     psi_l = psi_l, type = 0, ncores = 1, seed = 42)
#> 
#> Parametric bootstrap confidence intervals.
#> type = "bc"
#>              est     se  R    2.5%  97.5%
#> beta_1_1   0.500 0.0479 10  0.4513 0.5588
#> beta_2_1   0.000 0.0430 10 -0.0404 0.0828
#> beta_3_1   0.000 0.0719 10 -0.0528 0.1514
#> beta_1_2   0.000 0.0697 10 -0.0699 0.1102
#> beta_2_2   0.500 0.0546 10  0.4368 0.5753
#> beta_3_2   0.000 0.0371 10 -0.0130 0.0308
#> beta_1_3   0.000 0.0450 10 -0.0725 0.0522
#> beta_2_3   0.000 0.0411 10 -0.0400 0.0855
#> beta_3_3   0.500 0.0824 10  0.2898 0.5426
#> psi_1_1    0.001 0.0001 10  0.0009 0.0011
#> psi_2_2    0.001 0.0001 10  0.0009 0.0011
#> psi_3_3    0.001 0.0001 10  0.0009 0.0011
#> mu0_1_1    0.000 0.0083 10 -0.0086 0.0114
#> mu0_2_1    0.000 0.0103 10 -0.0108 0.0207
#> mu0_3_1    0.000 0.0132 10 -0.0279 0.0077
#> sigma0_1_1 0.001 0.0005 10  0.0004 0.0018
#> sigma0_2_2 0.001 0.0012 10  0.0004 0.0040
#> sigma0_3_3 0.001 0.0006 10  0.0006 0.0023
summary(pb, type = "bc")
#>              est     se  R    2.5%  97.5%
#> beta_1_1   0.500 0.0479 10  0.4513 0.5588
#> beta_2_1   0.000 0.0430 10 -0.0404 0.0828
#> beta_3_1   0.000 0.0719 10 -0.0528 0.1514
#> beta_1_2   0.000 0.0697 10 -0.0699 0.1102
#> beta_2_2   0.500 0.0546 10  0.4368 0.5753
#> beta_3_2   0.000 0.0371 10 -0.0130 0.0308
#> beta_1_3   0.000 0.0450 10 -0.0725 0.0522
#> beta_2_3   0.000 0.0411 10 -0.0400 0.0855
#> beta_3_3   0.500 0.0824 10  0.2898 0.5426
#> psi_1_1    0.001 0.0001 10  0.0009 0.0011
#> psi_2_2    0.001 0.0001 10  0.0009 0.0011
#> psi_3_3    0.001 0.0001 10  0.0009 0.0011
#> mu0_1_1    0.000 0.0083 10 -0.0086 0.0114
#> mu0_2_1    0.000 0.0103 10 -0.0108 0.0207
#> mu0_3_1    0.000 0.0132 10 -0.0279 0.0077
#> sigma0_1_1 0.001 0.0005 10  0.0004 0.0018
#> sigma0_2_2 0.001 0.0012 10  0.0004 0.0040
#> sigma0_3_3 0.001 0.0006 10  0.0006 0.0023
confint(pb, type = "bc")
#>                    2.5 %      97.5 %
#> beta_1_1    0.4512946903 0.558836344
#> beta_2_1   -0.0403928322 0.082848108
#> beta_3_1   -0.0528482046 0.151371459
#> beta_1_2   -0.0698588308 0.110227599
#> beta_2_2    0.4368115908 0.575332421
#> beta_3_2   -0.0130351289 0.030832505
#> beta_1_3   -0.0724607669 0.052155419
#> beta_2_3   -0.0399948946 0.085503162
#> beta_3_3    0.2897539685 0.542562597
#> psi_1_1     0.0008514919 0.001069493
#> psi_2_2     0.0008951854 0.001053960
#> psi_3_3     0.0009384488 0.001087969
#> mu0_1_1    -0.0086169076 0.011419101
#> mu0_2_1    -0.0107924456 0.020685178
#> mu0_3_1    -0.0278534096 0.007708978
#> sigma0_1_1  0.0003791673 0.001833989
#> sigma0_2_2  0.0003786557 0.003980955
#> sigma0_3_3  0.0005860189 0.002311270
# }